Ole, ole ole ole, Allay, Allay (READ DESC)
First thing's first: The Allay's movement is broken, and idk how to fix it. Once it's fixed, the Allay should be able to ONLY move to the item (not pick it up yet.)

This does work as the Allay goes near the item instead of wandering around, but it simply can't lower itself to the item, due to the broken movement.